Move CustomEntryType to model package
