+++ src/main/java/net/sf/jabref/imports/BibtexParser.java
@@ -41,10 +41,7 @@ import java.util.regex.Pattern;
 import net.sf.jabref.*;
 import net.sf.jabref.gui.GUIGlobals;
 import net.sf.jabref.logic.l10n.Localization;
-import net.sf.jabref.model.BibtexDatabase;
-import net.sf.jabref.model.BibtexEntry;
-import net.sf.jabref.model.BibtexEntryType;
-import net.sf.jabref.model.BibtexEntryTypes;
+import net.sf.jabref.model.*;
 import net.sf.jabref.util.Util;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
