+++ src/main/java/net/sf/jabref/model/BibtexEntryType.java
@@ -35,7 +35,6 @@ import java.util.Locale;
 import java.util.Set;
 import java.util.TreeMap;
 
-import net.sf.jabref.CustomEntryType;
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
 import net.sf.jabref.util.Util;
